backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.104.1",
    "pydantic[email]==2.5.0",
    "pydantic-settings==2.1.0",
    "uvicorn[standard]==0.24.0",
    "razorpay==1.3.0",
    "python-dotenv==1.0.0",
    "httpx==0.25.2",
    "requests==2.31.0",
    "python-json-logger==2.0.7",
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "pytest-cov==4.1.0",
    "black==23.12.0",
    "flake8==6.1.0",
    "pylint==3.0.3",
    "isort==5.13.2",
    "mypy==1.7.1",
    "types-requests==2.31.0.10",
]
